+++ webots_ros2_examples/launch/example.launch.py
@@ -21,11 +21,10 @@ import os
 import launch
 import launch_ros.actions
 
-import webots_ros2_core.webots_launcher
+from webots_ros2_core.utils import get_webots_home
 
 
 def generate_launch_description():
-    webotsHome = webots_ros2_core.webots_launcher.get_webots_home()
     arguments = ['--mode=realtime', '--world=' +
                  os.path.join(os.path.dirname(os.path.abspath(__file__)),
                               'worlds', 'ros_example.wbt')]
@@ -34,7 +33,7 @@ def generate_launch_description():
     controller = launch_ros.actions.Node(package='webots_ros2_examples',
                                          node_executable='example_controller',
                                          output='screen')
-    os.environ['LD_LIBRARY_PATH'] = (os.path.join(webotsHome, 'lib') + ':' +
+    os.environ['LD_LIBRARY_PATH'] = (os.path.join(get_webots_home(), 'lib') + ':' +
                                      os.environ.get('LD_LIBRARY_PATH'))
     return launch.LaunchDescription([
         webots,
